tiler/pyproject.toml
[project]
name = "stacnotator-tiler"
version = "0.1.0"
description = "Tile serving service for STACNotator"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.38.0",
    "gunicorn>=23.0.0",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    # Tile serving
    "titiler.core>=0.19.0",
    "titiler.mosaic>=0.19.0",
    "morecantile>=5.0.0",
    "rasterio>=1.3.0",
    "rio-tiler>=6.0.0",
    "numpy>=1.26.0",
    # STAC catalog search (for stats)
    "pystac-client>=0.8.0",
    # MPC signing
    "planetary-computer>=1.0.0",
]
